# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Point System API",
        "version": "1.0.0"
    },
    "paths": {
        "/pointsystem/public/v1/setuserpoint": {
            "post": {
                "tags": [
                    "SetUserPoint"
                ],
                "summary": "SetUserPoint. Set point both addition and reduction",
                "description": "SetUserPoint",
                "operationId": "setuserpoint",
                "parameters": [
                    {
                        "name": "point_type_id",
                        "in": "query",
                        "description": "Type of point",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Amount of point",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of point",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Incomplete parameter(s)",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Point process failed",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/pointsystem/public/v1/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login to authnticate user",
                "description": "Authenticate User",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Incomplete parameter(s)",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    }
}